优化：自增ID字段升级为bigint-角色

+++ sz-service/sz-service-admin/src/main/java/com/sz/admin/system/controller/SysRoleController.java
@@ -81,7 +81,7 @@ public class SysRoleController {
     @Operation(summary = "角色菜单信息查询")
     @SaCheckPermission(value = {"sys.role.setting_btn", "sys.role.update_btn"}, mode = SaMode.AND, orRole = GlobalConstant.SUPER_ROLE)
     @GetMapping("/menu")
-    public ApiResult findRoleMenuByRoleId(@NotZero @RequestParam Integer roleId) {
+    public ApiResult findRoleMenuByRoleId(@NotZero @RequestParam Long roleId) {
         return ApiPageResult.success(sysRoleMenuService.queryRoleMenu(roleId));
     }
